+++ app/src/main/java/ru/s4m1d/notes/app/desktop/client/app/runner/AppStarter.java
@@ -1,20 +1,19 @@
 package ru.s4m1d.notes.app.desktop.client.app.runner;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import ru.s4m1d.notes.app.desktop.client.business.context.SwingApplicationContext;
-import ru.s4m1d.notes.app.desktop.client.business.context.NotesContext;
-import ru.s4m1d.notes.app.desktop.client.business.context.WorkspaceContext;
-import ru.s4m1d.notes.app.desktop.client.business.task.MainFrameStateService;
-import ru.s4m1d.notes.app.desktop.client.business.task.context.ContextInitializationTask;
-import ru.s4m1d.notes.app.desktop.client.business.notes.NotesMonitoringServiceImpl;
-import ru.s4m1d.notes.app.desktop.client.business.notes.NotesBarUpdateServiceImpl;
-import ru.s4m1d.notes.app.desktop.client.business.task.workspace.WorkspaceObserver;
-import ru.s4m1d.notes.app.desktop.client.business.task.workspace.WorkspaceService;
-import ru.s4m1d.notes.app.desktop.client.business.task.workspace.WorkspaceWorker;
+import ru.s4m1d.notes.app.desktop.client.controller.context.SwingApplicationContext;
+import ru.s4m1d.notes.app.desktop.client.controller.context.NotesContext;
+import ru.s4m1d.notes.app.desktop.client.controller.context.WorkspaceContext;
+import ru.s4m1d.notes.app.desktop.client.controller.task.MainFrameStateService;
+import ru.s4m1d.notes.app.desktop.client.controller.task.context.ContextInitializationTask;
+import ru.s4m1d.notes.app.desktop.client.controller.notes.NotesMonitoringServiceImpl;
+import ru.s4m1d.notes.app.desktop.client.controller.notes.NotesBarUpdateServiceImpl;
+import ru.s4m1d.notes.app.desktop.client.controller.task.workspace.WorkspaceObserver;
+import ru.s4m1d.notes.app.desktop.client.controller.task.workspace.WorkspaceService;
+import ru.s4m1d.notes.app.desktop.client.controller.task.workspace.WorkspaceWorker;
 import ru.s4m1d.notes.app.desktop.client.components.*;
 import ru.s4m1d.notes.app.desktop.client.components.MenuBar;
 import ru.s4m1d.notes.app.desktop.client.components.choice.bar.NotesBarLayeredPane;
