--- controller-api/src/main/java/ru/s4m1d/notes/app/desktop/client/business/datamodel/content/Content.java
@@ -1,7 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.business.datamodel.content;
-
-import javax.swing.text.Document;
-
-public interface Content<T> {
-    public T get();
-}
