--- controller-api/src/main/java/ru/s4m1d/notes/app/desktop/client/business/task/NotesMonitoringService.java
@@ -1,5 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.business.task;
-
-public interface NotesMonitoringService {
-
-}
