--- controller-api/src/main/java/ru/s4m1d/notes/app/desktop/client/business/task/NotesPaneUpdateService.java
@@ -1,4 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.business.task;
-
-public interface NotesPaneUpdateService {
-}
