--- controller-api/src/main/java/ru/s4m1d/notes/app/desktop/client/business/task/NotesSearchService.java
@@ -1,9 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.business.task;
-
-import ru.s4m1d.notes.app.desktop.client.business.datamodel.note.Note;
-
-import java.util.List;
-
-public interface NotesSearchService {
-    public List<Note> searchForNotes();
-}
