+++ controller-api/src/main/java/ru/s4m1d/notes/app/desktop/client/controller/task/NotesMonitoringService.java
@@ -1,0 +1,5 @@
+package ru.s4m1d.notes.app.desktop.client.controller.task;
+
+public interface NotesMonitoringService {
+
+}
