+++ controller-api/src/main/java/ru/s4m1d/notes/app/desktop/client/controller/task/NotesPaneUpdateService.java
@@ -1,0 +1,4 @@
+package ru.s4m1d.notes.app.desktop.client.controller.task;
+
+public interface NotesPaneUpdateService {
+}
