--- controller-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/business/task/workspace/WorkspaceEvent.java
@@ -1,6 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.business.task.workspace;
-
-import ru.s4m1d.notes.app.desktop.client.core.Event;
-
-public class WorkspaceEvent implements Event {
-}
