--- controller-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/business/task/workspace/WorkspaceWorker.java
@@ -1,16 +1,0 @@
-package ru.s4m1d.notes.app.desktop.client.business.task.workspace;
-
-import lombok.AllArgsConstructor;
-
-import javax.swing.*;
-
-@AllArgsConstructor
-public class WorkspaceWorker extends SwingWorker<Void, Void> {
-    private WorkspaceService workspaceService;
-
-    @Override
-    protected Void doInBackground() throws Exception {
-        workspaceService.loadContent();
-        return null;
-    }
-}
