+++ controller-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/controller/notes/NotesMonitoringServiceObservable.java
@@ -1,0 +1,6 @@
+package ru.s4m1d.notes.app.desktop.client.controller.notes;
+
+import ru.s4m1d.notes.app.desktop.client.core.Observable;
+
+public class NotesMonitoringServiceObservable extends Observable {
+}
