+++ controller-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/controller/task/workspace/WorkspaceEvent.java
@@ -1,0 +1,6 @@
+package ru.s4m1d.notes.app.desktop.client.controller.task.workspace;
+
+import ru.s4m1d.notes.app.desktop.client.core.Event;
+
+public class WorkspaceEvent implements Event {
+}
